fix: uart适配的serials_buff_len存放的一直是个错误的值

+++ app/port/luat_uart_air101.c
@@ -17,7 +17,7 @@
 #define MAX_DEVICE_COUNT TLS_UART_MAX
 
 //存放串口设备句柄
-static uint8_t serials_buff_len[MAX_DEVICE_COUNT] ={TLS_UART_RX_BUF_SIZE};
+static uint32_t serials_buff_len[MAX_DEVICE_COUNT] ={TLS_UART_RX_BUF_SIZE};
 extern struct tls_uart_port uart_port[TLS_UART_MAX];
 
 typedef struct serials_timer_info {
